+++ src/llm-providers/mocked/MockedEchoLlmExecutionTools.ts
@@ -106,6 +106,5 @@ export class MockedEchoLlmExecutionTools implements LlmExecutionTools {
 
 /**
  * TODO: !!! Use here `seed`
- * TODO: [🕵️‍♀️] Maybe just remove
  * TODO: Allow in spaceTrim: nesting with > ${block(prompt.request)}, same as replace params
  */