+++ src/llm-providers/mocked/MockedFackedLlmExecutionTools.ts
@@ -130,5 +130,4 @@ export class MockedFackedLlmExecutionTools implements LlmExecutionTools {
 
 /**
  * TODO: !!! Use here `seed`
- * TODO: [🕵️‍♀️] Maybe just remove
  */