+++ src/llm-providers/utils/cache/CacheLlmToolsOptions.ts
@@ -1,0 +1,11 @@
+import type { PromptbookStorage } from '../../../storage/_common/PrompbookStorage';
+import type { CacheItem } from './CacheItem';
+
+export type CacheLlmToolsOptions = {
+    /**
+     * Total cost of the execution
+     *
+     * @default MemoryStorage
+     */
+    storage: PromptbookStorage<CacheItem>;
+};