+++ src/storage/files-storage/FilesStorageOptions.ts
@@ -1,0 +1,11 @@
+import { string_folder_path } from '../../types/typeAliases';
+
+/**
+ * @@@
+ */
+export type FilesStorageOptions = {
+    /**
+     * @@@
+     */
+    folderPath: string_folder_path;
+};