+++ src/storage/files-storage/utils/nameToSubfolderPath.ts
@@ -1,0 +1,11 @@
+import type { string_name } from '../../typeAliases';
+
+
+/**
+ * @@@
+ *
+ * @private for `FilesStorage`
+ */
+export function nameToSubfolderPath(name: string_name): Array<string> {
+    return [name.substr(0, 1).toLowerCase(), name.substr(1, 1).toLowerCase()];
+}