+++ src/storage/local-storage/LocalStorage.ts
@@ -1,0 +1,10 @@
+import { makePromptbookStorageFromWebStorage } from '../utils/makePromptbookStorageFromWebStorage';
+
+/**
+ * @@@
+ */
+export const LocalStorage = makePromptbookStorageFromWebStorage(localStorage);
+
+/**
+ * Export through `@promptbook/browser`
+ */