+++ src/storage/local-storage/SessionStorage.ts
@@ -1,0 +1,10 @@
+import { makePromptbookStorageFromWebStorage } from '../utils/makePromptbookStorageFromWebStorage';
+
+/**
+ * @@@
+ */
+export const SessionStorage = makePromptbookStorageFromWebStorage(sessionStorage);
+
+/**
+ * Export through `@promptbook/browser`
+ */